# perf: temporary cache for consequent calls of CDeterministicMNList for historical blocks