--- a/src/evo/deterministicmns.cpp
+++ b/src/evo/deterministicmns.cpp
@@ -778,6 +778,13 @@ CDeterministicMNList CDeterministicMNManager::GetListForBlockInternal(gsl::not_n
     for (const auto& diffIndex : listDiffIndexes) {
         const auto& diff = mnListDiffsCache.at(diffIndex->GetBlockHash());
         snapshot.ApplyDiff(diffIndex, diff);
+        if (snapshot.GetHeight() % 32 == 0) {
+            // Add this temporary mini-snapshot to cache.
+            // This extra cached mn-list helps to improve performance of GetListForBlock
+            // for close blocks, because in the worst cases each of them requires to retrieve
+            // and apply up to 575 diffs
+            mnListsCache.emplace(snapshot.GetBlockHash(), snapshot);
+        }
     }
 
     if (tipIndex) {