@knst 2025-06-24
@UdjinM6 @PastaPastaPasta 

at first, I haven't planned to get this one merged, because I assumed that it improves only `UndoBlock` for several blocks; which maybe useful for reconsiderblock; but actually it improves RPC performance for `protx diff`, especially for multiple calls for blocks that are close to each other. Also `protx list` is improved, if it's called multiple times for blocks that are close to each other in chain.