@UdjinM6 2025-08-15
1. We should probably disable this extra cache while node is still reindexing e.g
```diff
diff --git a/src/evo/deterministicmns.cpp b/src/evo/deterministicmns.cpp
index 921a119642..3f2dab1d3e 100644
--- a/src/evo/deterministicmns.cpp
+++ b/src/evo/deterministicmns.cpp
@@ -16,6 +16,7 @@
 #include <consensus/validation.h>
 #include <deploymentstatus.h>
 #include <messagesigner.h>
+#include <node/blockstorage.h>
 #include <script/standard.h>
 #include <stats/client.h>
 #include <uint256.h>
@@ -26,6 +27,8 @@
 #include <optional>
 #include <memory>
 
+using node::fReindex;
+
 static const std::string DB_LIST_SNAPSHOT = "dmn_S3";
 static const std::string DB_LIST_DIFF = "dmn_D3";
 
@@ -780,7 +783,7 @@ CDeterministicMNList CDeterministicMNManager::GetListForBlockInternal(gsl::not_n
     for (const auto& diffIndex : listDiffIndexes) {
         const auto& diff = mnListDiffsCache.at(diffIndex->GetBlockHash());
         snapshot.ApplyDiff(diffIndex, diff);
-        if (snapshot.GetHeight() % 32 == 0) {
+        if (!fReindex && (snapshot.GetHeight() % 32 == 0)) {
             // Add this temporary mini-snapshot to cache.
             // This extra cached mn-list helps to improve performance of GetListForBlock
             // for close blocks, because in the worst cases each of them requires to retrieve
```

2.
> 1. Consider Making Cache Interval Configurable
>    - The hardcoded value 32 could be a named constant or configurable parameter
>    static constexpr int MINI_SNAPSHOT_INTERVAL = 32;
>    if (snapshot.GetHeight() % MINI_SNAPSHOT_INTERVAL == 0) {

Agree

3.
> 1. Add Test Coverage
>    
>    * The PR lacks unit tests verifying the caching behavior
>    * Consider adding tests to ensure:
>      
>      * Snapshots are created at correct intervals
>      * Cache hits improve performance
>      * Cache cleanup doesn't remove actively used snapshots

Adding unit tests here feels like an overkill tbh